+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -2709,8 +2709,12 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
     }
     String branch = gitProperties.getProperty("git.branch");
     if (version.endsWith("-SNAPSHOT") || !branch.equals("master")) {
+      // Sometimes the branch is HEAD, which is not informative.
+      // How does that happen, and how can I fix it?
       version += ", branch " + branch;
-      version += ", commit " + gitProperties.getProperty("git.commit.id");
+      // For brevity, only date but not time of day.
+      version += ", " + gitProperties.getProperty("git.commit.time").substring(0, 10);
+      version += ", commit " + gitProperties.getProperty("git.commit.id.abbrev");
     }
     return version;
   }