+++ docs/CHANGELOG.md
@@ -19,6 +19,13 @@ Instead of overriding `isRelevant()`, a type factory implementation should
 override `isRelevantImpl()`.  Clients should continue to call `isRelevant()`;
 never call `isRelevantImpl()`.
 
+Methods that now return a `boolean` rather than `void`:
+ * `commonAssignmentCheck()`
+ * `checkArrayInitialization()`
+ * `checkLock()`
+ * `checkLockOfThisOrTree()`
+ * `ensureExpressionIsEffectivelyFinal()`
+
 Methods that now return `AnnotationMirrorSet` instead of `Set<? extends AnnotationMirror>`:
  * `getTopAnnotations()`
  * `getBottomAnnotations()`
@@ -30,6 +37,8 @@ Renamed `BaseTypeVisitor.checkExtendsImplements()` to `checkExtendsAndImplements
 Class `FieldInvariants`:
  * constructor now takes an `AnnotatedTypeFactory`
  * `isSuperInvariant()` no longer takes an `AnnotatedTypeFactory`
+
+Removed methods that have been deprecated for over two years.
 
 **Closed issues:**
 
@@ -1061,7 +1070,7 @@ All CFGVisualizeLauncher command-line arguments now start with `--` instead of `
 
 **Implementation details:**
 
-commonAssignmentCheck() now takes an additional argument.  Type system
+`commonAssignmentCheck()` now takes an additional argument.  Type system
 authors must update their overriding implementations.
 
 Renamed methods: